pyproject.toml
[project]
name = "parted"
version = "0.1.0"
description = "A christian music app"
authors = [
    {name = "Gopar",email = "gopardaniel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "django-cotton (>=1.5.2,<2.0.0)",
    "django-htmx (>=1.22.0,<2.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "django-allauth[socialaccount] (>=65.4.1,<66.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
rich = "^13.9.4"
mypy = "^1.15.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.3"}
faker = "^36.1.1"
django-browser-reload = "^1.18.0"
djade = "^1.3.2"
django-debug-toolbar = "^5.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
addopts = "-vv"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "config/settings",
    "migrations",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"


[tool.mypy]
plugins = ["mypy_django_plugin.main"]
# Enables strict type checking
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true

# Allow untyped decorators to avoid unnecessary errors
disallow_untyped_decorators = false

# Ensure mypy follows imports correctly
follow_imports = "normal"
ignore_missing_imports = true

# Enable specific checks
warn_redundant_casts = true
warn_unreachable = true

# Exclude migrations from checking as they are autogenerated
exclude = ["migrations", "__pycache__", "tests", "manage.py"]

[tool.django-stubs]
django_settings_module = "config.settings.local"
